requirements.txt
# LangChain Core
langchain
langchain-Core

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
# langchain-google-gemini
# google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment variable management
python-dotenv

# Machine learning utilities
numpy
scikit-learn